requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
python-pptx>=1.0.0
